+++ Frends.AzureDataLake.DeleteContainer/Frends.AzureDataLake.DeleteContainer/Definitions/Options.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 
 namespace Frends.AzureDataLake.DeleteContainer.Definitions;
 
@@ -13,4 +14,20 @@ public class Options
     /// <example>false</example>
     [DefaultValue(false)]
     public bool ThrowErrorIfContainerDoesNotExist { get; set; }
+
+    /// <summary>
+    /// Throw an exception if the task encounters an error.
+    /// When false, the task returns a Result with Success = false and Error details instead.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Optional custom error message to use when ThrowErrorOnFailure is true.
+    /// </summary>
+    /// <example></example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
 }